requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cycler==0.10.0
Django==2.2.6
django-ckeditor==5.7.1
django-cleanup==4.0.0
django-crispy-forms==1.7.2
django-js-asset==1.2.2
Flask==1.1.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
mysqlclient==1.4.4
numpy==1.17.2
pandas==0.25.1
passlib==1.7.1
Pillow==6.2.0
pyparsing==2.4.2
PyQt5==5.13.1
PyQt5-sip==12.7.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.3.10
sqlparse==0.3.0
urllib3==1.25.6
Werkzeug==0.16.0
WTForms==2.2.1
